soft/requirements_import.txt
# Para rodar import_csv_supabase.py
supabase>=2.0.0
python-dotenv>=1.0.0
